+++ src/view/_loader.py
@@ -3,14 +3,32 @@ from __future__ import annotations
 import os
 import runpy
 import warnings
+from dataclasses import _MISSING_TYPE, Field
 from pathlib import Path
-from typing import TYPE_CHECKING, Iterable, get_args
+from typing import TYPE_CHECKING, Iterable, TypedDict, get_args, get_type_hints, NamedTuple
+
+from pydantic.fields import ModelField
 
 from ._logging import Internal
 from ._util import set_load
 from .exceptions import InvalidBodyError, LoaderWarning
-from .routing import Method, Route, RouteInput, _NoDefault
+from .routing import BodyParam, Method, Route, RouteInput, _NoDefault
 from .typing import Any, RouteInputDict, TypeInfo, ValueType
+
+ExtNotRequired = None
+try:
+    from typing import NotRequired
+except ImportError:
+    NotRequired = None
+    from typing_extensions import NotRequired as ExtNotRequired
+
+_NOT_REQUIRED_TYPES = []
+
+if ExtNotRequired:
+    _NOT_REQUIRED_TYPES.append(ExtNotRequired)
+
+if NotRequired:
+    _NOT_REQUIRED_TYPES.append(NotRequired)
 
 if TYPE_CHECKING:
     from .app import ViewApp
@@ -26,6 +44,7 @@ TYPECODE_FLOAT = 4
 TYPECODE_DICT = 5
 TYPECODE_NONE = 6
 TYPECODE_CLASS = 7
+TYPECODE_CLASSTYPES = 8
 
 
 _BASIC_CODES = {
@@ -43,10 +62,49 @@ Type info should contain three things:
     - Type Code
     - Type Object (only set when using a __view_body__ object)
     - Children (i.e. the `int` part of dict[str, int])
+    - Default (only set when typecode is TYPECODE_CLASSTYPES)
 
 This can be formatted as so:
     [(union1_tc, None, []), (union2_tc, None, [(type_tc, obj, [])])]
 """
+
+
+class _ViewNotRequired:
+    __VIEW_NOREQ__ = 1
+
+
+def _format_body(vbody_types: dict) -> list[TypeInfo]:
+    if not isinstance(vbody_types, dict):
+        raise InvalidBodyError(
+            f"__view_body__ should return a dict, not {type(vbody_types)}",  # noqa
+        )
+
+    vbody_final = {}
+    vbody_defaults = {}
+
+    for k, raw_v in vbody_types.items():
+        if not isinstance(k, str):
+            raise InvalidBodyError(
+                f"all keys returned by __view_body__ should be strings, not {type(k)}"  # noqa
+            )
+
+        default = _NoDefault
+        v = raw_v.types if isinstance(raw_v, BodyParam) else raw_v
+
+        if isinstance(raw_v, BodyParam):
+            default = raw_v.default
+
+        if getattr(raw_v, "__origin__", None) in _NOT_REQUIRED_TYPES:
+            v = get_args(raw_v)
+
+        iter_v = v if isinstance(v, (tuple, list)) else (v,)
+        vbody_final[k] = _build_type_codes(iter_v)
+        vbody_defaults[k] = default
+
+    return [
+        (TYPECODE_CLASSTYPES, k, v, vbody_defaults[k])
+        for k, v in vbody_final.items()
+    ]
 
 
 def _build_type_codes(inp: Iterable[type[ValueType]]) -> list[TypeInfo]:
@@ -62,17 +120,87 @@ def _build_type_codes(inp: Iterable[type[ValueType]]) -> list[TypeInfo]:
             codes.append((type_code, None, []))
             continue
 
+        if TypedDict in getattr(tp, "__orig_bases__", []):
+            body = get_type_hints(tp)
+
+            class _Transport:
+                @staticmethod
+                def __view_body__():
+                    return body
+
+                @staticmethod
+                def __view_construct__(**kwargs):
+                    return kwargs
+
+            codes.append(
+                (
+                    TYPECODE_CLASS,
+                    _Transport,
+                    _format_body(body),
+                ),
+            )
+            continue
+
+        if NamedTuple in getattr(tp, "__orig_bases__", []):
+            defaults = tp._field_defaults  # type: ignore
+            tps = {}
+            hints = get_type_hints(tp)
+
+            for k, v in hints.items():
+                if k in defaults:
+                    tps[k] = BodyParam(v, defaults[k])
+                else:
+                    tps[k] = v
+
+            codes.append((TYPECODE_CLASS, tp, _format_body(tps)))
+            continue
+
+        dataclass_fields: dict[str, Field] | None = getattr(
+            tp, "__dataclass_fields__", None
+        )
+
+        if dataclass_fields:
+            tps = {}
+            for k, v in dataclass_fields.items():
+                if isinstance(v.default, _MISSING_TYPE) and (
+                    isinstance(v.default_factory, _MISSING_TYPE)
+                ):
+                    tps[k] = v.type
+                else:
+                    default = (
+                        v.default
+                        if not isinstance(v.default, _MISSING_TYPE)
+                        else v.default_factory
+                    )
+                    tps[k] = BodyParam(v.type, default)
+
+            codes.append((TYPECODE_CLASS, tp, _format_body(tps)))
+            continue
+
+        pydantic_fields: dict[str, ModelField] | None = getattr(
+            tp, "__fields__", None
+        )
+        if pydantic_fields:
+            tps = {}
+
+            for k, v in pydantic_fields.items():
+                if not v.default:
+                    tps[k] = v.type_
+                else:
+                    tps[k] = BodyParam(v.type_, v.default)
+
+            codes.append((TYPECODE_CLASS, tp, _format_body(tps)))
+            continue
+
         vbody = getattr(tp, "__view_body__", None)
         if vbody:
-            raise NotImplementedError
-            """
             if callable(vbody):
                 vbody_types = vbody()
             else:
                 vbody_types = vbody
 
-            codes.append((TYPECODE_CLASS, tp, _build_type_codes(vbody_types)))
-            """
+            codes.append((TYPECODE_CLASS, tp, _format_body(vbody_types)))
+            continue
 
         origin = getattr(tp, "__origin__", None)  # typing.GenericAlias
 
@@ -191,13 +319,16 @@ def load_fs(app: ViewApp, target_dir: Path):
                 else:
                     path_obj = Path(path)
                     stripped = list(
-                        path_obj.parts[len(target_dir.parts) :]
-                    )  # noqa
+                        path_obj.parts[len(target_dir.parts) :]  # noqa
+                    )
                     if stripped[-1] == "index.py":
                         stripped.pop(len(stripped) - 1)
 
                     stripped_obj = Path(*stripped)
-                    stripped_path = str(stripped_obj).rsplit(".", maxsplit=1)[0]
+                    stripped_path = str(stripped_obj).rsplit(
+                        ".",
+                        maxsplit=1,
+                    )[0]
                     x.path = "/" + stripped_path
 
             for x in current_routes: