+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated `__all__` for `routing.py`
 - Added `view.Response` and `view.HTML`
 - Fixed `__view_result__`
+- Added support for `__view_body__` and `__view_construct__`
+- Added support for Pydantic, `NamedTuple`, and dataclasses for type validation
 
 ## [1.0.0-alpha4] - 2023-09-10
 - Added type validation (without support for `__view_body__`)